change migration

+++ console/migrations/m231223_082050_create_income_table.php
@@ -15,7 +15,7 @@ class m231223_082050_create_income_table extends Migration
         $this->createTable('income', [
             'id' => $this->primaryKey(),
             'product_id' => $this->integer(),
-            'amount' => $this->decimal(10, 2)->notNull(),
+            'amount' => $this->integer()->notNull(),
             'created_at' => $this->timestamp()->defaultExpression('CURRENT_TIMESTAMP'),
             'updated_at' => $this->timestamp()->defaultExpression('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'),
             'deleted_at' => $this->timestamp()->null(),
